Create a module for ast interfaces and rename files
