+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './astProcessor'
+export * from './ast-processor'
 export * from './interfaces'
 export * from './utils'
