+++ src/interfaces/ast-mappings/document.ts
