+++ src/interfaces/process-result.ts
@@ -1,4 +1,4 @@
-import { Feature } from './feature'
+import { Feature } from './ast-mappings'
 
 export interface ProcessResult {
     features: Feature[]
