+++ src/interfaces/prototyper.ts
@@ -1,4 +1,4 @@
-import { Feature } from './feature'
+import { Feature } from './ast-mappings'
 
 export interface Prototyper {
   generate(features: Feature[]): Promise<string[]>
